Convert to typescript the mute and unmute slash commands files
